\name{Revo.home}
\alias{Revo.home}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Return the Revolution R Home Directory}
\description{Return the Revolution R home directory.
}
\usage{
Revo.home(component = "home")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
      \item{component}{One of the strings \code{"home"}, \code{"doc"}, or
      \code{"licenses"}. }
}
\details{
    \code{Revo.home} returns the path to the Revolution R installation
    directory, or its doc or licenses subdirectory. 
}
\value{
a character string containing the path to the
Revolution R installation directory. 
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{utilities}
\keyword{documentation}
