\name{RevoInfo}
\alias{RevoLicense}
\alias{forum}
\alias{readme}
\alias{revo}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Revolution R Information Functions}
\description{Functions giving information about Revolution R.
}
\usage{
RevoLicense()
revo()
forum()
readme()
}
%- maybe also 'usage' for other objects documented here.

\details{
    \code{RevoLicense} displays the Revolution R license. \code{revo} 
    displays the Revolution Analytics
    home page. \code{forum} displays the Revolution R User Forum. 
    \code{readme} displays the Revolution R release notes.
}
\value{
    For \code{RevoLicense}, \code{revo},
    \code{forum}, and \code{readme}, \code{NULL} is returned invisibly.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{utilities}
\keyword{documentation}
